#2 [feat] 심화과제: TopTabBar에 따른 화면 이동
