+++ 2nd assignment/2nd assignment/Home/HomeViewController/Dummy/ParamountViewController.swift
@@ -1,0 +1,29 @@
+//
+//  ParamountViewController.swift
+//  2nd assignment
+//
+//  Created by 왕정빈 on 2024/05/03.
+//
+
+import UIKit
+
+class ParamountViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        view.backgroundColor = .lightGray
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
