+++ 2nd assignment/2nd assignment/Home/Model/HomeSectionType.swift
@@ -1,0 +1,12 @@
+//
+//  HomeSectionType.swift
+//  2nd assignment
+//
+//  Created by 왕정빈 on 2024/05/03.
+//
+
+import Foundation
+
+enum Section: Int, CaseIterable {
+    case Main, MustSeen, PopularLive, FreeContent, AD, MagicContent
+}
